Added a sign up activity
