+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="app/src/main/res/drawable/arrow_downward-48px.xml" value="0.196" />
+        <entry key="app/src/main/res/drawable/arrow_right_alt-48px.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/arrow_upward-48px.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/bottom_gray_border.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/bottom_gray_border_blue.xml" value="0.348" />
@@ -16,7 +17,7 @@
         <entry key="app/src/main/res/drawable/near_me-48px.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/person-48px.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/small_corner.xml" value="0.348" />
-        <entry key="app/src/main/res/layout/activity_login.xml" value="0.3233836830085052" />
+        <entry key="app/src/main/res/layout/activity_login.xml" value="0.22071186904366144" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.3161953727506427" />
         <entry key="app/src/main/res/layout/fragment_create.xml" value="0.37244873046875" />
         <entry key="app/src/main/res/layout/fragment_home.xml" value="0.1" />
