+++ app/src/main/res/drawable/icon_right_arrow.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <path
+      android:fillColor="@android:color/white"
+      android:pathData="M28.05,35.9 L25.9,33.8 34.3,25.4H8V22.4H34.3L25.85,13.95L28,11.85L40.05,23.9Z"/>
+</vector>
